+++ lib/assets/javascripts/new-dashboard/pages/Data/Subscriptions.vue
@@ -1,6 +1,29 @@
 <template>
   <section class="subscriptions-section">
     <div class="container grid">
+      <div class="full-width">
+        <SectionTitle class="grid-cell" :showActionButton="true">
+          <template slot="icon">
+            <img src="../../assets/icons/subscriptions/subscriptions_tick-icon.svg" width="18" height="20" />
+          </template>
+          <template slot="title">
+            <VisualizationsTitle :defaultTitle="$t('DataPage.tabs.subscriptions')"/>
+          </template>
+          <template slot="actionButton">
+            <router-link :to="{ name: 'spatial-data-catalog' }"  exact>
+              <button class="button is-primary">{{$t('Subscriptions.new')}}</button>
+            </router-link>
+          </template>
+        </SectionTitle>
+      </div>
+      <div v-if="loading" class="u-flex u-flex__direction--column u-flex__align--center u-width--100 u-mt--120">
+        <span class="loading u-mb--12">
+          <img svg-inline src="../../assets/icons/common/loading.svg" class="loading__svg"/>
+        </span>
+        <span class="text is-txtSoftGrey is-caption">
+          Loading your subscriptions…
+        </span>
+      </div>
       <div class="u-width--100" v-if="!loading">
         <div v-if="subscriptions.length === 0" class="grid-cell grid-cell--col12">
           <EmptyState
@@ -14,31 +37,15 @@
           </router-link>
         </div>
         <template v-else>
-          <SectionTitle class="grid-cell" :showActionButton="true">
-            <template slot="icon">
-              <img src="../../assets/icons/subscriptions/subscriptions_tick-icon.svg" width="18" height="20" />
-            </template>
-
-            <template slot="title">
-              <VisualizationsTitle :defaultTitle="$t(`DataPage.tabs.yourSubscriptions`)"/>
-            </template>
-
-            <template slot="dropdownButton">
-              <SettingsDropdown>
-                <img svg-inline src="../../assets/icons/common/filter.svg">
-              </SettingsDropdown>
-            </template>
-
-            <template slot="actionButton">
-              <button class="button is-primary">{{$t('Subscriptions.new')}}</button>
-            </template>
-          </SectionTitle>
           <ul>
-            <div class="subscription-item u-flex" v-for="subscription in subscriptions" :key="subscription.slug">
+            <div class="subscription-item u-flex" v-for="subscription in subscriptionsByPage" :key="subscription.slug">
               <DatasetListItem :dataset="subscription"></DatasetListItem>
-              <DatasetListItemExtra></DatasetListItemExtra>
+              <DatasetListItemExtra :dataset="subscription"></DatasetListItemExtra>
             </div>
           </ul>
+          <div class="u-mt--48 u-flex u-flex__justify--center">
+            <Pager :count="count" :currentPage="currentPage" @goToPage="goToPage"></Pager>
+          </div>
         </template>
       </div>
     </div>
@@ -52,8 +59,9 @@ import EmptyState from 'new-dashboard/components/States/EmptyState';
 import SectionTitle from 'new-dashboard/components/SectionTitle';
 import VisualizationsTitle from 'new-dashboard/components/VisualizationsTitle';
 import SettingsDropdown from 'new-dashboard/components/Settings/Settings';
-import DatasetListItem from '@carto/common-ui/do-catalog/src/components/catalogSearch/DatasetListItem';
+import DatasetListItem from 'new-dashboard/components/Catalog/browser/DatasetListItem';
 import DatasetListItemExtra from 'new-dashboard/components/Subscriptions/DatasetListItemExtra';
+import Pager from 'new-dashboard/components/Catalog/browser/Pager';
 
 export default {
   name: 'SubscriptionsPage',
@@ -63,33 +71,92 @@ export default {
     VisualizationsTitle,
     SettingsDropdown,
     DatasetListItem,
-    DatasetListItemExtra
+    DatasetListItemExtra,
+    Pager
   },
   data () {
     return {
-      loading: true
+      loading: true,
+      currentPage: 0,
+      id_interval: null
     };
   },
   computed: {
     ...mapState({
-      subscriptions: state => state.doCatalog.subscriptionsList
-    })
+      subscriptions: state => state.catalog.subscriptionsList
+    }),
+    pageSize () {
+      return process.env.VUE_APP_PAGE_SIZE || 10;
+    },
+    count () {
+      return this.subscriptions.filter(s => s.slug).length;
+    },
+    subscriptionsByPage () {
+      return this.subscriptions.filter(s => s.slug).slice(
+        this.currentPage * this.pageSize,
+        (this.currentPage + 1) * this.pageSize
+      );
+    },
+    isAnySubscriptionSyncing () {
+      return this.subscriptions && this.subscriptions.find(s => s.sync_status === 'syncing');
+    }
   },
   async mounted () {
-    await this.$store.dispatch('doCatalog/fetchSubscriptionsList', true);
-    this.loading = false;
+    this.loading = true;
+    await this.$store.dispatch('catalog/fetchSubscriptionsList');
+    await this.fetchSubscriptionsListDetail();
   },
-  beforeDestroy () {},
-  beforeRouteUpdate (to, from, next) {},
-  methods: {}
+  methods: {
+    async fetchSubscriptionsListDetail () {
+      this.loading = true;
+      window.scrollTo(0, 0);
+      await this.$store.dispatch('catalog/fetchSubscriptionsDetailsList', this.subscriptions.map(s => s.id));
+      this.loading = false;
+    },
+    goToPage (pageNum) {
+      this.currentPage = pageNum;
+      window.scrollTo(0, 0);
+    }
+  },
+  watch: {
+    isAnySubscriptionSyncing: {
+      immediate: true,
+      handler () {
+        clearInterval(this.id_interval);
+        if (this.isAnySubscriptionSyncing) {
+          this.id_interval = setInterval(async () => {
+            await this.$store.dispatch('catalog/fetchSubscriptionsList', true);
+          }, 5000);
+        }
+      }
+    }
+  },
+  destroyed () {
+    clearInterval(this.id_interval);
+  }
 };
 </script>
 
 <style scoped lang="scss">
 @import 'new-dashboard/styles/variables';
 
+.full-width {
+  width: 100%;
+}
+
 .subscriptions-section {
+  min-height: 640px;
   margin-top: 64px;
+
+  &__filter {
+    justify-content: space-between;
+    height: 168px;
+
+    &--dropdown {
+      position: relative;
+    }
+  }
+
   .empty-state {
     margin: 20vh 0 58px;
     /deep/ h6 {
@@ -111,8 +178,33 @@ export default {
       border-bottom: 1px solid $neutral--300;
     }
     .list-item {
+      flex: 1 1 100%;
+      display: flex;
+      flex-direction: column;
       &:hover {
         background-color: transparent;
+      }
+      /deep/ {
+        .extra {
+          flex: 1 1 100%;
+          >div {
+            display: flex;
+            align-items: flex-end;
+          }
+        }
+      }
+    }
+  }
+  .loading {
+    &__svg {
+      width: 40px;
+      stroke: $blue--500;
+      g {
+        stroke-width: 2px;
+        circle {
+          stroke:#36434A;
+          stroke-opacity: 0.25;
+        }
       }
     }
   }